Add the Calculator Library class library
