+++ Calculator.nbcraig/Calculator.nbcraig.sln
@@ -1,9 +1,11 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
-VisualStudioVersion = 17.13.35818.85 d17.13
+VisualStudioVersion = 17.13.35818.85
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Calculator.nbcraig", "Calculator.nbcraig\Calculator.nbcraig.csproj", "{FBED0CD7-57A1-4A2E-96DC-73968106414D}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CalculatorLibrary", "CalculatorLibrary\CalculatorLibrary.csproj", "{812C89D9-F8B6-480A-ABCF-A5C4159EDA5C}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -15,6 +17,10 @@ Global
 		{FBED0CD7-57A1-4A2E-96DC-73968106414D}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{FBED0CD7-57A1-4A2E-96DC-73968106414D}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{FBED0CD7-57A1-4A2E-96DC-73968106414D}.Release|Any CPU.Build.0 = Release|Any CPU
+		{812C89D9-F8B6-480A-ABCF-A5C4159EDA5C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{812C89D9-F8B6-480A-ABCF-A5C4159EDA5C}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{812C89D9-F8B6-480A-ABCF-A5C4159EDA5C}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{812C89D9-F8B6-480A-ABCF-A5C4159EDA5C}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
